requirements.txt
azure-functions
fastapi
uvicorn[standard]
nest_asyncio